fix: fix toggle button for deleted config being reversed
Fixes #1952
